Refactor device fonts to favor code reuse
