+++ components/micropython/port/src/omv/img/include/font_device.h
@@ -1,0 +1,9 @@
+/*
+ * This file is overridden by the device-specific font.
+ * 
+ * static uint8_t unicode[]
+ * static uint8_t unicode_wide[]
+ * font_config
+ * font_config_wide
+ *
+ */
